+++ src/scripting/lua_color.cpp
@@ -16,7 +16,7 @@
 #include "scripting/lua_color.hpp"
 #include "scripting/lua_common.hpp"
 #include "scripting/push_check.hpp"
-#include "lua/lauxlib.h"
+#include "lua/wrapper_lauxlib.h"
 #include "log.hpp"
 #include "game_config.hpp"
