+++ src/scripting/lua_stringx.cpp
@@ -23,7 +23,7 @@
 
 #include <boost/algorithm/string/trim.hpp>
 
-#include "lua/lauxlib.h"
+#include "lua/wrapper_lauxlib.h"
 
 namespace lua_stringx {
