+++ src/scripting/lua_team.cpp
@@ -26,7 +26,7 @@
 
 #include <string>
 
-#include "lua/lauxlib.h"
+#include "lua/wrapper_lauxlib.h"
 
 /**
  * Implementation for a lua reference to a team,
