+++ src/scripting/lua_terrainmap.cpp
@@ -27,7 +27,7 @@
 #include "game_board.hpp"
 #include "play_controller.hpp"
 
-#include "lua/lauxlib.h"
+#include "lua/wrapper_lauxlib.h"
 
 static lg::log_domain log_scripting_lua("scripting/lua");
 #define LOG_LUA LOG_STREAM(info, log_scripting_lua)
