+++ src/scripting/lua_unit.cpp
@@ -31,7 +31,7 @@
 #include "game_version.hpp"
 #include "deprecation.hpp"
 
-#include "lua/lauxlib.h"
+#include "lua/wrapper_lauxlib.h"
 
 static lg::log_domain log_scripting_lua("scripting/lua");
 #define LOG_LUA LOG_STREAM(info, log_scripting_lua)
