+++ src/scripting/lua_unit_attacks.cpp
@@ -24,7 +24,7 @@
 #include "units/attack_type.hpp"
 #include "utils/const_clone.hpp"
 
-#include "lua/lauxlib.h"
+#include "lua/wrapper_lauxlib.h"
 
 #include <type_traits>
