+++ src/scripting/lua_widget_methods.cpp
@@ -51,7 +51,7 @@
 #include <utility>
 #include <vector>
 
-#include "lua/lauxlib.h"
+#include "lua/wrapper_lauxlib.h"
 
 static lg::log_domain log_scripting_lua("scripting/lua");
 #define ERR_LUA LOG_STREAM(err, log_scripting_lua)
